+++ protocols/valve/dota2/webapi/src/main/java/com/ibasco/agql/protocols/valve/dota2/webapi/pojos/Dota2MatchHistoryInfo.java
@@ -34,9 +34,11 @@ public class Dota2MatchHistoryInfo {
     @SerializedName("match_id")
     @Expose
     private long matchId;
+
     @SerializedName("match_seq_num")
     @Expose
-    private int matchSeqNum;
+    private long matchSeqNum;
+
     @SerializedName("start_time")
     @Expose
     private int startTime;
@@ -77,7 +79,7 @@ public class Dota2MatchHistoryInfo {
      *
      * @return The matchSeqNum
      */
-    public int getMatchSeqNum() {
+    public long getMatchSeqNum() {
         return matchSeqNum;
     }
 
@@ -87,7 +89,7 @@ public class Dota2MatchHistoryInfo {
      * @param matchSeqNum
      *         The match_seq_num
      */
-    public void setMatchSeqNum(int matchSeqNum) {
+    public void setMatchSeqNum(long matchSeqNum) {
         this.matchSeqNum = matchSeqNum;
     }
